add framework de mapping + backend dos formularios

+++ src/main/java/br/com/ucommex/resourceserver/feedback/dto/CustomerFeedbackDTO.java
@@ -7,5 +7,5 @@ public class CustomerFeedbackDTO {
     private String cpf;
     private String infoAtendimento;
     private String infoAtendimentoOutraLoja;
-    private long nps;
+    private int nps;
 }
